Function index

AIBECS functions

AIBECS.OCIM1.load — Function.

load

Returns wet3d, grd, and T (in that order) from FigShare repository.

source

AIBECS.GridTools.vector_of_depths — Function.

vector_of_depths(wet3D, grid)

Returns the vector of depths of the center of wet boxes.

source

AIBECS.GridTools.number_of_wet_boxes — Function.

number_of_wet_boxes(wet3D)

Returns the number of wet grid boxes.

source

AIBECS.GridTools.indices_of_wet_boxes — Function.

indices_of_wet_boxes(wet3D)

Returns the vector of the indices of wet grid boxes.

source

Missing docstring.

Missing docstring for empty_parameter_table. Check Documenter's build log for details.

AIBECS.add_parameter! — Function.

add_parameter!(t::DataFrame, args...; kwargs...)

Adds a parameter to the parameters table t. If keyword argument optimizable = false, then observation mean and variance are set to NaN. Otherwise, these are set to keyword arguments mean_obs (and variance_obs) if supplied, or to quantity (and its square), after converting it to the preferred unit and stripping it of said unit if not. Example: TODO Note for future edit of the docs: Don't repeat yourself between add and new param functions

source

AIBECS.initialize_Parameters_type — Function.

initialize_Parameters_type(t, PName="Parameters")

Generate the type called after PName and all its functionality with it. It is recommended to use upper camel case for PName as for all user-defined Julia types. PName defaults to "Parameters".

For example, use

julia> initialize_Parameters_type(t) # creates Parameters

or

julia> initialize_Parameters_type(t, "MyPara") # creates MyPara

source

AIBECS.state_function_and_Jacobian — Function.

F, ∇ₓF = state_function_and_Jacobian(Ts, Gs, nb)

Returns the state function F and its jacobian, ∇ₓF.

source

DiffEqBase.SteadyStateProblem — Type.

SteadyStateProblem(F, ∇ₓF, x, p)

Returns the SteadyStateProblem defined by F(x,p)=0.

source

DiffEqBase.solve — Function.

solve(prob::DiffEqBase.AbstractSteadyStateProblem,
      alg::CTKAlg;
      nrm=norm,
      τstop=1e12*365*24*60*60,
      preprint="",
      maxItNewton=50)

Solves prob using the modified C.T.Kelley Shamanskii algorithm.

source